+++ lib/greenhouse_api/resources/candidates.rb
@@ -5,11 +5,11 @@ require 'sorbet-runtime'
 
 module GreenhouseApi
   module Resources
-    class Candidates < Client
+    class Candidates < BaseClient
       extend T::Sig
 
       sig { params(params: T.nilable(T::Hash[String, T.any(Integer, String)])).returns(Response) }
-      def self.list_candidates(params = {})
+      def self.list_all(params)
         list_many('candidates', params)
       end
     end
